+++ libavcodec/hevcdsp_template.c
@@ -560,7 +560,7 @@ static void FUNC(put_hevc_pel_bi_pixels)(uint8_t *_dst, ptrdiff_t _dststride, ui
 }
 
 static void FUNC(put_hevc_pel_uni_w_pixels)(uint8_t *_dst, ptrdiff_t _dststride, uint8_t *_src, ptrdiff_t _srcstride,
-                                          int width, int height, intptr_t mx, intptr_t my, int denom, int wx, int ox)
+                                          int width, int height, intptr_t mx, intptr_t my, int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     pixel *src          = (pixel *)_src;
@@ -881,7 +881,7 @@ static void FUNC(put_hevc_qpel_bi_hv)(uint8_t *_dst, ptrdiff_t _dststride, uint8
 static void FUNC(put_hevc_qpel_uni_w_h)(uint8_t *_dst,  ptrdiff_t _dststride,
                                       uint8_t *_src, ptrdiff_t _srcstride,
                                       int width, int height, intptr_t mx, intptr_t my,
-                                      int denom, int wx, int ox)
+                                      int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     pixel        *src       = (pixel*)_src;
@@ -936,7 +936,7 @@ static void FUNC(put_hevc_qpel_bi_w_h)(uint8_t *_dst, ptrdiff_t _dststride, uint
 static void FUNC(put_hevc_qpel_uni_w_v)(uint8_t *_dst,  ptrdiff_t _dststride,
                                       uint8_t *_src, ptrdiff_t _srcstride,
                                       int width, int height, intptr_t mx, intptr_t my,
-                                      int denom, int wx, int ox)
+                                      int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     pixel        *src       = (pixel*)_src;
@@ -992,7 +992,7 @@ static void FUNC(put_hevc_qpel_bi_w_v)(uint8_t *_dst, ptrdiff_t _dststride, uint
 static void FUNC(put_hevc_qpel_uni_w_hv)(uint8_t *_dst,  ptrdiff_t _dststride,
                                        uint8_t *_src, ptrdiff_t _srcstride,
                                        int width, int height, intptr_t mx, intptr_t my,
-                                       int denom, int wx, int ox)
+                                       int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     const int8_t *filter;
@@ -1323,7 +1323,7 @@ static void FUNC(put_hevc_epel_bi_hv)(uint8_t *_dst, ptrdiff_t _dststride, uint8
 }
 
 static void FUNC(put_hevc_epel_uni_w_h)(uint8_t *_dst, ptrdiff_t _dststride, uint8_t *_src, ptrdiff_t _srcstride,
-                                      int width, int height, intptr_t mx, intptr_t my, int denom, int wx, int ox)
+                                      int width, int height, intptr_t mx, intptr_t my, int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     pixel *src = (pixel *)_src;
@@ -1375,7 +1375,7 @@ static void FUNC(put_hevc_epel_bi_w_h)(uint8_t *_dst, ptrdiff_t _dststride, uint
 }
 
 static void FUNC(put_hevc_epel_uni_w_v)(uint8_t *_dst, ptrdiff_t _dststride, uint8_t *_src, ptrdiff_t _srcstride,
-                                      int width, int height, intptr_t mx, intptr_t my, int denom, int wx, int ox)
+                                      int width, int height, intptr_t mx, intptr_t my, int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     pixel *src = (pixel *)_src;
@@ -1427,7 +1427,7 @@ static void FUNC(put_hevc_epel_bi_w_v)(uint8_t *_dst, ptrdiff_t _dststride, uint
 }
 
 static void FUNC(put_hevc_epel_uni_w_hv)(uint8_t *_dst, ptrdiff_t _dststride, uint8_t *_src, ptrdiff_t _srcstride,
-                                       int width, int height, intptr_t mx, intptr_t my, int denom, int wx, int ox)
+                                       int width, int height, intptr_t mx, intptr_t my, int16_t denom, int16_t wx, int16_t ox)
 {
     int x, y;
     pixel *src = (pixel *)_src;
